fix: rusqlite 0.40 usize + sha2/hmac API compat
- hex crate for sha256 LowerHex removal in sha2 0.11
- hmac::KeyInit import for HmacSha256::new_from_slice
- Option<usize> -> Option<i64> for rusqlite 0.40 ToSql/FromSql
- FailoverQueueItem sort_index: usize -> i64

+++ src-tauri/src/database/dao/failover.rs
@@ -13,7 +13,7 @@ use serde::{Deserialize, Serialize};
 pub struct FailoverQueueItem {
     pub provider_id: String,
     pub provider_name: String,
-    pub sort_index: Option<usize>,
+    pub sort_index: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub provider_notes: Option<String>,
 }
